+++ src/emmental/learner.py
@@ -2,6 +2,7 @@
 import collections
 import copy
 import logging
+import time
 from collections import defaultdict
 from typing import Dict, List, Optional, Union
 
@@ -96,7 +97,7 @@ class EmmentalLearner(object):
                 **optimizer_config[f"{opt}_config"],
             )
         elif isinstance(opt, optim.Optimizer):  # type: ignore
-            optimizer = opt(parameters)
+            optimizer = opt(parameters)  # type: ignore
         else:
             raise ValueError(f"Unrecognized optimizer option '{opt}'")
 
@@ -492,6 +493,8 @@ class EmmentalLearner(object):
           dataloaders: A list of dataloaders used to learn the model.
         """
         # Generate the list of dataloaders for learning process
+        start_time = time.time()
+
         train_split = Meta.config["learner_config"]["train_split"]
         if isinstance(train_split, str):
             train_split = [train_split]
@@ -520,6 +523,28 @@ class EmmentalLearner(object):
         self._set_optimizer(model)
         # Set up lr_scheduler
         self._set_lr_scheduler(model)
+
+        if Meta.config["learner_config"]["fp16"]:
+            try:
+                from apex import amp  # type: ignore
+            except ImportError:
+                raise ImportError(
+                    "Please install apex from https://www.github.com/nvidia/apex to "
+                    "use fp16 training."
+                )
+            logger.info(
+                f"Modeling training with 16-bit (mixed) precision "
+                f"and {Meta.config['learner_config']['fp16_opt_level']} opt level."
+            )
+            model, self.optimizer = amp.initialize(
+                model,
+                self.optimizer,
+                opt_level=Meta.config["learner_config"]["fp16_opt_level"],
+            )
+
+        # Multi-gpu training (after apex fp16 initialization)
+        if Meta.config["model_config"]["dataparallel"]:
+            model._to_dataparallel()
 
         # Set to training mode
         model.train()
@@ -584,7 +609,11 @@ class EmmentalLearner(object):
                     )
 
                     # Perform backward pass to calculate gradients
-                    loss.backward()  # type: ignore
+                    if Meta.config["learner_config"]["fp16"]:
+                        with amp.scale_loss(loss, self.optimizer) as scaled_loss:
+                            scaled_loss.backward()
+                    else:
+                        loss.backward()  # type: ignore
 
                 if (total_batch_num + 1) % Meta.config["learner_config"][
                     "optimizer_config"
@@ -594,12 +623,20 @@ class EmmentalLearner(object):
                 ):
                     # Clip gradient norm
                     if Meta.config["learner_config"]["optimizer_config"]["grad_clip"]:
-                        torch.nn.utils.clip_grad_norm_(
-                            model.parameters(),
-                            Meta.config["learner_config"]["optimizer_config"][
-                                "grad_clip"
-                            ],
-                        )
+                        if Meta.config["learner_config"]["fp16"]:
+                            torch.nn.utils.clip_grad_norm_(
+                                amp.master_params(self.optimizer),
+                                Meta.config["learner_config"]["optimizer_config"][
+                                    "grad_clip"
+                                ],
+                            )
+                        else:
+                            torch.nn.utils.clip_grad_norm_(
+                                model.parameters(),
+                                Meta.config["learner_config"]["optimizer_config"][
+                                    "grad_clip"
+                                ],
+                            )
 
                     # Update the parameters
                     self.optimizer.step()
@@ -615,3 +652,4 @@ class EmmentalLearner(object):
                 self._update_lr_scheduler(model, total_batch_num, self.metrics)
 
         model = self.logging_manager.close(model)
+        logger.info(f"Total learning time: {time.time() - start_time} seconds.")